requirements.txt
apache-airflow
apache-airflow-providers-apache-spark
pyspark
SQLAlchemy
kafka-python
cassandra-driver